requirements.txt
streamlit
easyocr
numpy
opencv-python-headless
psycopg2-binary
python-dotenv
groq
requests
face-recognition
dlib